Implementation of User Interface class, optimization of methods, change gitignore

+++ src/com/rumos/bank/administration/ADM.java
@@ -9,29 +9,29 @@ import com.rumos.bank.administration.models.DebitCard;
 
 public class ADM {
 
-	private static Long accountNumber = 100L;
-	private static Long clientNumber = 0L;
-	private static Long creditCardNumber = 10L;
-	private static Long debitCardNumber = 10L;
+	private static int accountNumber = 100;
+	private static int clientNumber = 0;
+	private static int creditCardNumber = 10;
+	private static int debitCardNumber = 10;
 	public static ArrayList<Client> clients = new ArrayList<>();
 	public static ArrayList<Account> accounts = new ArrayList<>();
 	public static ArrayList<DebitCard> debitCards = new ArrayList<>();
 	public static ArrayList<CreditCard> creditCards = new ArrayList<>();
 	
 	
-	public static Long accountNumber() {
+	public static int accountNumber() {
 		return ++accountNumber;
 	}
 	
-	public static Long clientNumber() {
+	public static int clientNumber() {
 		return ++clientNumber;
 	}
 	
-	public static Long creditCardNumber() {
+	public static int creditCardNumber() {
 		return ++creditCardNumber;
 	}
 	
-	public static Long debitCardNumber() {
+	public static int debitCardNumber() {
 		return ++debitCardNumber;
 	}
 }
